+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] - 2023-03-21
+### Added
+
+- support image ``![image-title](image-url)``
+
+### Fixed
+
+- Line-height
+
+
 ## [0.0.7] - 2023-03-20
 
 ### Changed
@@ -39,11 +49,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- 
+- support image ``![image-title](image-url)``
 
 ### Fixed
 
-- Line-height
+- 
 
 ### Changed